+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ### 1.3.1 (Next)
 - Properly model nested struct fields as optional when documented as such.
 - Improve required parameter validation.
-- Fix various issues occurring during `check` step when `version` parameter is omitted.
+- Fix issues occurring during `check` step when `version` parameter is omitted.
 
 ### 1.3.0
 - Support Vault Kubernetes authentication method.